changes to IndoorNormalClassifier
The IndoorNormalClassifier should as of now always be part of the
internal meshing pipeline. All faces are now colored in standard
lightgrey. Instead of coloring, the IndoorNormalClassifier now creates
for each region a string "label" being one of "horizontalupper",
"horizontallower", "vertical" or "unknown". In further steps, these
region labels can be evalutated.