change `#include "A/X.hpp"` to `#include <lvr/A/X.hpp>` in headers
Actually `#include "../A/X.hpp"` is the preferred way to include here,
but gcc doesn't shorten such include pathes and stacks them recursively
e.g. including "lvr/reconstruction/../geometry/../io/../io/PointBuffer.hpp"

In theory this is no problem, but in practice
(1.) it becomes harder to read error messages
(2.) Windows has rather short file path limits and this mechanism is known
to produce compile failures on Windows because of the increased path length.

Headers in the same section are included via `#include "X.hpp"` to
keep the sections as self-contained as possible and reduce file lookups.