collect installation include dirs of the separate projects
These will be added to lvrConfig.cmake later on.
Adding them to the INTERFACE of the library targets
as implemented before fails to capture header-only libraries.
INTERFACE libraries could be used to resolve this,
but wouldn't leave a way to use only the lvr headers
(without the library). Also adding "proper" interfaces
to the libs requires to transform every find_packaged
external dependency into a target with proper interfaces.
Most configs do not provide these yet.