Improve `calcVertexLocalNeighborhood()` slightly
The function was improved by using a better suited method from
`BaseMesh`. This way we avoid some unnecessary heap allocations
which improves performance a bit. A very unscientific benchmark
shows that the new version of `calcVertexHeightDiff` is twice
as fast.

These changes do not change any results of those algorithms.