Fix `calcVertexLocalNeighborhood`: avoid duplicate neighbors
Before, many vertices were pushed multiple times due to a small bug
in the implementation. The `neighborsOut` vector usually contained
somewhere from twice to thrice as many elements as necessary, meaning
that every vertex handle in this vector was (on average) two or three
times in that vector.

This changes has two effects:
- The result of `calcVertexRoughness` is slightly different! As this
  function averages over a value attached to each vertex, having some
  vertices multiple times or only once makes a difference. However,
  this difference isn't huge. Apart from that, I think that the new
  value is the correct one while the old one was incorrect due to
  said bug.
- Since functions using `calcVertexLocalNeighborhood` get a smaller
  vector now, those functions are faster. My unscientific benchmark
  shows an improvement of around 80% (e.g. 16s --> 9,5s).