Slightly improve local neighborhood search and add visitor
The `calcVertexLocalNeighborhood` is slightly faster now (only
by around 15%). There is also a new method `visitLocalNeighborhoodOfVertex`
which uses a visitor instead of pushing all neighbors onto a vector.
The former function uses the latter.

The visit function is now used in height-diff and roughness
calculations. It hardly improves performance, but makes the code a
bit shorter.