Add `containsFEV()` method to BaseMesh
This is particular useful for OpenMP usage. Of course, it's nice
to use the `mesh.vertices()` iterator, but OpenMP can't make
a parallel for loop out of that. Thus, you need to manually iterate
over all vertices by iterating over all possible handles. But you
need to be able to check whether a particular vertex/... exists.
That's what these methods are for.