Precompile Logger and Monitor
Implementing the Logger and Monitor functions using the spdlog and
spdmon libraries in a .cpp file removes the need to install spdlog and
spdmon and therefore prevents conflicts with other installations of
these libraries. spdlog::spdlog and spdmon::spdmon can be privately
linked against the lvr2core, lvr2_static and lvr2 targets.